Fix tests

+++ server/routes/tasks.js
@@ -18,7 +18,6 @@ export default (app) => {
     .get('/tasks', { name: 'tasks', preValidation: app.authenticate }, async (req, reply) => {
       const { query, user: { id } } = req;
       logApp('.get tasks req.query %O', req.query);
-      logApp('req.cookies %O', req.cookies);
 
       const tasksQuery = app.objection.models.task.query()
         .withGraphJoined('[creator, executor, status, labels]');
@@ -79,6 +78,9 @@ export default (app) => {
           }], { relate: ['labels'] });
           logApp('task.$relatedQuery to labels %O', await task.$relatedQuery('labels'));
         });
+        const fromDbTask = await app.objection.models.task
+          .query().findOne({ name: req.body.data.name });
+        logApp('task from DB =>', fromDbTask);
 
         req.flash('info', i18next.t('flash.task.create.success'));
         reply.redirect(app.reverse('tasks'));
@@ -98,6 +100,8 @@ export default (app) => {
           task, users, statuses, labels, errors: err.data,
         });
         return reply.code(422);
+      } finally {
+        logApp('End of handling in "createTask"');
       }
     })
 
@@ -135,8 +139,17 @@ export default (app) => {
 
     .delete('/tasks/:id', {
       name: 'deleteTask',
-      preValidation: app.auth([app.checkIfUserCreatedTask, app.authenticate]),
+      preValidation: app.authenticate,
     }, async (req, reply) => {
+      // logApp('req =>', req)
+      const { creatorId } = await app.objection.models.task.query().findById(req.params.id);
+      logApp('creatorId %O', creatorId);
+      logApp('req.user.id %O', req.user.id);
+      if (req.user.id !== creatorId) {
+        logApp('User is not creator of task');
+        req.flash('error', i18next.t('flash.task.authError'));
+        return reply.redirect('/tasks');
+      }
       await app.objection.models.task.query().deleteById(req.params.id);
       req.flash('info', i18next.t('flash.task.delete.success'));
       reply.redirect(app.reverse('tasks'));
